Use render prop instead of component prop for Route components
#### Issue
For the following code in `src/components/SignUp.js`
```
auth.doCreateUserWithEmailAndPassword(email, passwordOne)
  .then(authUser => {
    db.doCreateUser(authUser.user.uid, username, email)
      .then(() => {
        this.setState({ ...INITIAL_STATE });
        history.push(routes.HOME);
      })
  })
```
presently the following warning is shown in console while doing a new sign-up:
```
Warning: Can't call setState (or forceUpdate) on an unmounted component. This is a no-op, but it indicates a memory leak in your application.
```

#### Root cause
As per the react-router docs page:
```
"If you provide an inline function to the component prop, you would create a new component every render. This results in the existing component unmounting and the new component mounting instead of just updating the existing component."
```
In the `src/components/App.js` file, we are currently using component prop in the Route component:
```
<Route exact path={routes.SIGN_UP} component={() => <SignUpPage />} />
```
As a result of creation a new component on every render, the SignUpPage component gets remounted on every re-render.
As a consequence, by the time db.doCreateUser() promise is resolved, the previous SignUpPage component gets unmounted and gets replaced with a new component causing the previously mentioned warning "Can't call setState on an unmounted component".

#### Fix
As per the react-router docs page:
```
"The render prop allows for convenient inline rendering and wrapping without the undesired remounting explained above."
```
So, any of the following two changes will fix the issue.
```
<Route exact path={routes.SIGN_UP} component={SignUpPage} />
```
OR
```
<Route exact path={routes.SIGN_UP} render={() => <SignUpPage />} />
```
This PR chooses the later solution since it involves minimal change in the code.
Similar changes are also done to all other Route components used in the code at different places.