Correct English language and add Part numbers
Corrected some minor grammar errors.
Improved the style.
Removed some less common wording.
Simplified the style.
Added Part 1 to the title
Added part numbers to the links at the bottom -- the actual parts themselves are not numbered elsewhere, but if this proposal is accepted then I will number the other parts accordingly
I corrected  "uses a [function as a child]" but it could also be "uses [functions as children]". At this moment, I am not sure which is better in this context.